chore: move static assets generation to its own module chore: move some utilities to utils.ts
