+++ src/types.ts
@@ -94,7 +94,7 @@ export interface VitePWAOptions {
    */
   include: string | string[] | undefined
   /**
-   * By default the icons listed on `manifest` options will be included
+   * By default the icons listed on `manifest` option will be included
    * on the service worker *precache* if present under Vite's `publicDir`
    * option directory.
    */
